Move withBFTNode to CardanoNode and add tests
This also asserts the "right" cardano-version is available as we want to
avoid false-positives in our integration tests -> see note in code.
