+++ hydra-cluster/hydra-cluster.cabal
@@ -190,6 +190,7 @@ test-suite integration
     Paths_hydra_cluster
     Spec
     Test.CardanoClusterSpec
+    Test.CardanoNodeSpec
     Test.DirectChainSpec
     Test.EndToEndSpec
     Test.GeneratorSpec
@@ -207,6 +208,7 @@ test-suite integration
     , cardano-ledger-shelley
     , cardano-ledger-shelley-ma
     , containers
+    , directory
     , filepath
     , hedgehog-quickcheck
     , hspec
@@ -266,7 +268,5 @@ benchmark bench-e2e
     , strict-containers
     , time
 
-  build-tool-depends:
-    hydra-node:hydra-node -any
-
+  build-tool-depends: hydra-node:hydra-node -any
   ghc-options:        -threaded -rtsopts
