+++ hydra-cluster/src/CardanoNode.hs
@@ -8,27 +8,27 @@ module CardanoNode where
 import Hydra.Prelude
 
 import Control.Retry (constantDelay, limitRetriesByCumulativeDelay, retrying)
-import Control.Tracer (
-  Tracer,
-  traceWith,
- )
+import Control.Tracer (Tracer, traceWith)
 import Data.Aeson ((.=))
 import qualified Data.Aeson as Aeson
 import qualified Data.HashMap.Strict as HM
 import Data.Time.Clock.POSIX (posixSecondsToUTCTime, utcTimeToPOSIXSeconds)
 import Hydra.Cardano.Api (AsType (AsPaymentKey), PaymentKey, SigningKey, VerificationKey, generateSigningKey, getVerificationKey)
-import System.Directory (doesFileExist, removeFile)
 import Hydra.Cluster.Util (readConfigFile)
+import System.Directory (createDirectoryIfMissing, doesFileExist, removeFile)
 import System.Exit (ExitCode (..))
 import System.FilePath ((<.>), (</>))
+import System.Posix (ownerReadMode, setFileMode)
 import System.Process (
   CreateProcess (..),
   StdStream (UseHandle),
   proc,
   readCreateProcessWithExitCode,
+  readProcess,
   withCreateProcess,
  )
 import Test.Hydra.Prelude
+import Test.Network.Ports (randomUnusedTCPPort)
 
 type Port = Int
 
@@ -97,6 +97,80 @@ data PortsConfig = PortsConfig
   deriving stock (Show, Eq, Generic)
   deriving anyclass (ToJSON, FromJSON)
 
+getCardanoNodeVersion :: IO String
+getCardanoNodeVersion =
+  readProcess "cardano-node" ["--version"] ""
+
+-- | Start a cardano-node in BFT mode using the config from config/ and
+-- credentials from config/credentials/ using given 'nodeId'. NOTE: This means
+-- that nodeId should only be 1,2 or 3 and that only the faucet receives
+-- 'initialFunds'. Use 'seedFromFaucet' to distribute funds other wallets.
+withBFTNode ::
+  Tracer IO NodeLog ->
+  CardanoNodeConfig ->
+  (RunningNode -> IO ()) ->
+  IO ()
+withBFTNode tracer cfg action = do
+  createDirectoryIfMissing False (stateDirectory cfg)
+
+  [dlgCert, signKey, vrfKey, kesKey, opCert] <-
+    forM
+      [ dlgCertFilename nid
+      , signKeyFilename nid
+      , vrfKeyFilename nid
+      , kesKeyFilename nid
+      , opCertFilename nid
+      ]
+      (copyCredential (stateDirectory cfg))
+
+  let args =
+        defaultCardanoNodeArgs
+          { nodeDlgCertFile = Just dlgCert
+          , nodeSignKeyFile = Just signKey
+          , nodeVrfKeyFile = Just vrfKey
+          , nodeKesKeyFile = Just kesKey
+          , nodeOpCertFile = Just opCert
+          , nodePort = Just (ours (ports cfg))
+          }
+
+  readConfigFile "cardano-node.json"
+    >>= writeFileBS
+      (stateDirectory cfg </> nodeConfigFile args)
+
+  readConfigFile "genesis-byron.json"
+    >>= writeFileBS
+      (stateDirectory cfg </> nodeByronGenesisFile args)
+
+  readConfigFile "genesis-shelley.json"
+    >>= writeFileBS
+      (stateDirectory cfg </> nodeShelleyGenesisFile args)
+
+  readConfigFile "genesis-alonzo.json"
+    >>= writeFileBS
+      (stateDirectory cfg </> nodeAlonzoGenesisFile args)
+
+  withCardanoNode tracer cfg args $ \rn@(RunningNode _ socket) -> do
+    traceWith tracer $ MsgNodeStarting cfg
+    waitForSocket rn
+    traceWith tracer $ MsgSocketIsReady socket
+    action rn
+ where
+  dlgCertFilename i = "delegation-cert.00" <> show (i - 1) <> ".json"
+  signKeyFilename i = "delegate-keys.00" <> show (i - 1) <> ".key"
+  vrfKeyFilename i = "delegate" <> show i <> ".vrf.skey"
+  kesKeyFilename i = "delegate" <> show i <> ".kes.skey"
+  opCertFilename i = "opcert" <> show i <> ".cert"
+
+  copyCredential parentDir file = do
+    bs <- readConfigFile ("credentials" </> file)
+    let destination = parentDir </> file
+    unlessM (doesFileExist destination) $
+      writeFileBS destination bs
+    setFileMode destination ownerReadMode
+    pure destination
+
+  nid = nodeId cfg
+
 withCardanoNode ::
   Tracer IO NodeLog ->
   CardanoNodeConfig ->
@@ -126,6 +200,25 @@ withCardanoNode tr cfg@CardanoNodeConfig{stateDirectory, nodeId} args action = d
 
   socketFile = stateDirectory </> nodeSocket args
 
+newNodeConfig :: FilePath -> IO CardanoNodeConfig
+newNodeConfig stateDirectory = do
+  nodePort <- randomUnusedTCPPort
+  systemStart <- initSystemStart
+  pure $
+    CardanoNodeConfig
+      { nodeId = 1
+      , stateDirectory
+      , systemStart
+      , ports = PortsConfig nodePort []
+      }
+
+-- | Wait for the node socket file to become available.
+waitForSocket :: RunningNode -> IO ()
+waitForSocket node@(RunningNode _ socket) = do
+  unlessM (doesFileExist socket) $ do
+    threadDelay 0.1
+    waitForSocket node
+
 -- | Generate command-line arguments for launching @cardano-node@.
 cardanoNodeProcess :: Maybe FilePath -> CardanoNodeArgs -> CreateProcess
 cardanoNodeProcess cwd args = (proc "cardano-node" strArgs){cwd}
@@ -149,6 +242,12 @@ cardanoNodeProcess cwd args = (proc "cardano-node" strArgs){cwd}
   opt arg = \case
     Nothing -> []
     Just val -> [arg, val]
+
+-- | Initialize the system start time to now (modulo a small offset needed to
+-- give time to the system to bootstrap correctly).
+initSystemStart :: IO UTCTime
+initSystemStart = do
+  addUTCTime 1 <$> getCurrentTime
 
 -- | Re-generate configuration and genesis files with fresh system start times.
 refreshSystemStart :: CardanoNodeConfig -> CardanoNodeArgs -> IO ()
@@ -277,6 +376,7 @@ data NodeLog
   | MsgCLIStatus Text Text
   | MsgCLIRetry Text
   | MsgCLIRetryResult Text Int
+  | MsgNodeStarting CardanoNodeConfig
   | MsgSocketIsReady FilePath
   deriving stock (Eq, Show, Generic)
   deriving anyclass (ToJSON, FromJSON)
